+++ tests/test_gauss_iod.rs
@@ -23,7 +23,7 @@ fn run_iod(
     env_state: &mut Outfit,
     traj_set: &mut TrajectorySet,
     traj_number: &ObjectNumber,
-) -> Result<(Option<GaussResult>, f64), OutfitError> {
+) -> Result<(GaussResult, f64), OutfitError> {
     let obs = traj_set.get_mut(traj_number).unwrap();
     let mut rng = StdRng::seed_from_u64(42_u64); // seed for reproducibility
 
@@ -71,8 +71,7 @@ fn test_gauss_iod() {
         mean_anomaly: 0.44069989140091426,
     });
 
-    let best_orbit_unwrapped = best_orbit.unwrap();
-    let orbit = best_orbit_unwrapped.get_orbit();
+    let orbit = best_orbit.get_orbit();
 
     assert!(approx_equal(&expected_orbit, orbit, test_epsilon));
     assert_relative_eq!(
@@ -99,8 +98,7 @@ fn test_gauss_iod() {
         mean_anomaly: 4.85070383704545,
     });
 
-    let best_orbit_unwrapped = best_orbit.unwrap();
-    let orbit = best_orbit_unwrapped.get_orbit();
+    let orbit = best_orbit.get_orbit();
 
     assert!(approx_equal(&expected_orbit, orbit, test_epsilon));
     assert_relative_eq!(
@@ -127,8 +125,7 @@ fn test_gauss_iod() {
         mean_anomaly: 4.9466622638827324,
     });
 
-    let best_orbit_unwrapped = best_orbit.unwrap();
-    let orbit = best_orbit_unwrapped.get_orbit();
+    let orbit = best_orbit.get_orbit();
 
     assert!(approx_equal(&expected_orbit, orbit, test_epsilon));
     assert_relative_eq!(